Remove code smell; make NamespaceType a subclass of ObjectType
Addresses PR review feedback flagging `respond_to?(:namespace?) && namespace?`
in `HasIndices#index` as a Java-instanceof-style code smell.

Instead of feature-testing inside the generic `index` method, promote namespace
semantics into the type system: `NamespaceType < ObjectType` overrides
`namespace?` to return `true`, overrides `index` to raise, and moves
`auto_wire_namespace_subfields` onto itself. `ObjectType#namespace?` now simply
returns `false`, and the smell is removed from `HasIndices#index`.
