+++ elasticgraph-schema_definition/lib/elastic_graph/schema_definition/factory.rb
@@ -24,6 +24,7 @@ require "elastic_graph/schema_definition/schema_elements/graphql_sdl_enumerator"
 require "elastic_graph/schema_definition/schema_elements/input_field"
 require "elastic_graph/schema_definition/schema_elements/input_type"
 require "elastic_graph/schema_definition/schema_elements/interface_type"
+require "elastic_graph/schema_definition/schema_elements/namespace_type"
 require "elastic_graph/schema_definition/schema_elements/object_type"
 require "elastic_graph/schema_definition/schema_elements/relationship"
 require "elastic_graph/schema_definition/schema_elements/scalar_type"
@@ -223,6 +224,11 @@ module ElasticGraph
         end
       end
       @@interface_type_new = prevent_non_factory_instantiation_of(SchemaElements::InterfaceType)
+
+      def new_namespace_type(name, &block)
+        @@namespace_type_new.call(@state, name.to_s, &(_ = block))
+      end
+      @@namespace_type_new = prevent_non_factory_instantiation_of(SchemaElements::NamespaceType)
 
       def new_object_type(name)
         @@object_type_new.call(@state, name.to_s) do |object_type|
