+++ elasticgraph-schema_definition/lib/elastic_graph/schema_definition/mixins/has_indices.rb
@@ -68,10 +68,6 @@ module ElasticGraph
               "Indices must be configured during initial type definition."
           end
 
-          if respond_to?(:namespace?) && namespace?
-            raise Errors::SchemaError, "`#{self.name}` cannot be both an indexed type and a namespace type."
-          end
-
           if @own_index_def
             raise Errors::SchemaError, "Cannot define multiple indices on `#{self.name}`. " \
               "Only one index per type is supported. An index named `#{@own_index_def.name}` has already been defined."
