+++ elasticgraph-schema_definition/sig/elastic_graph/schema_definition/api.rbs
@@ -67,7 +67,7 @@ module ElasticGraph
       def json_schema_strictness: (?allow_omitted_fields: bool, ?allow_extra_fields: bool) -> void
       def raw_sdl: (::String) -> void
       def object_type: (::String) { (SchemaElements::ObjectType) -> void } -> void
-      def namespace_type: (::String) ?{ (SchemaElements::ObjectType) -> void } -> void
+      def namespace_type: (::String) ?{ (SchemaElements::NamespaceType) -> void } -> void
       def interface_type: (::String) { (SchemaElements::InterfaceType) -> void } -> void
       def enum_type: (::String) { (SchemaElements::EnumType) -> void } -> void
       def union_type: (::String) { (SchemaElements::UnionType) -> void } -> void
